1. Added ImageComparisonTools object for keeping static methods. 2. Added Unit test for it. 3. refactored code according to the Coverall and BetterCodeHub results.
